added player data loader

+++ SavePlayerData.java
@@ -1,6 +1,9 @@
 package finalProject;
 
+import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
+import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.io.Serializable;
 import java.util.ArrayList;
@@ -8,7 +11,6 @@ import java.util.ArrayList;
 public class SavePlayerData	extends PlayerData implements Serializable {
 	private static final long serialVersionUID = 1L;
     private static final String filepath = "C:\\Temp\\";
-    int[] saveSlotStatus = {0,0,0,0,0};
     
     private String name;
 	private int[] score = new int[2];
@@ -23,27 +25,47 @@ public class SavePlayerData	extends PlayerData implements Serializable {
 		this.difficulty = difficulty;
     }
     
-    public void writePlayerData(PlayerData player) {
-    	int openSaveSlots = 0;
-    	for(int i = 0; i < 5; ++i) {
-    		if(saveSlotStatus[i] == 0) {
-    			openSaveSlots = i;
+    public int writePlayerData(PlayerData player, boolean overwrite) {
+    	Exception e = null;
+    	File f = new File(filepath + player.getName() + ".txt");
+    	if(f.exists() && !f.isDirectory() && !overwrite) { 
+    	    System.out.println("File already exists");
+    	    return 2; // failed - file already exists error 
+    	}
+    	else {
+    		try {
+    			FileOutputStream playerData = new FileOutputStream(filepath + player.getName() + ".txt"); // all player data files will be the player's name
+    			System.out.println("File found.");
+    			ObjectOutputStream savePlayerData = new ObjectOutputStream(playerData);
+    			
+    			savePlayerData.writeObject(player);
+    			savePlayerData.close();			
+    			
+    			System.out.println("Player data saved successfully.");
+    		}catch(Exception e1) {
+    			e = e1;
+    			e.printStackTrace();
+    		}
+    		if(e == null) {
+    			return 1; // passed
+    		}
+    		else {
+    			return 0; // failed
     		}
     	}
-		try {
-			FileOutputStream playerData = new FileOutputStream(filepath + player.getName() + ".txt");
-			ObjectOutputStream savePlayerData = new ObjectOutputStream(playerData);
-			
-			savePlayerData.writeObject(player);
-			savePlayerData.close();			
-			
-			System.out.println("Player data saved successfully.");
-		}catch(Exception e) {
-			e.printStackTrace();
-		}
 	}	
 	
-	public void readPlayerData(PlayerData player) {
-		// TODO create player data reader and implement reader into GUI
+	public PlayerData readPlayerData(String inputName) {
+		PlayerData player = null;
+		try {
+			FileInputStream playerData = new FileInputStream(filepath + inputName + ".txt");
+			ObjectInputStream readPlayerData = new ObjectInputStream(playerData);
+			
+			player = (PlayerData) readPlayerData.readObject();
+			readPlayerData.close();
+		}catch(Exception e){
+			e.printStackTrace();
+		}
+		return player;
 	}
 }
